% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{display_constants}
\alias{display_constants}
\title{This is a example documentation
It looks like this
make sure you have the 'export' to make the function reachable from R (look at what happens at  R/RcppExports.R)}
\usage{
display_constants()
}
\description{
This is a example documentation
It looks like this
make sure you have the 'export' to make the function reachable from R (look at what happens at  R/RcppExports.R)
}
